requirements.txt
amqp==5.2.0 ; python_version >= "3.11" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.11" and python_version < "4.0"
billiard==4.2.0 ; python_version >= "3.11" and python_version < "4.0"
celery==5.4.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.6.2 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.16.0 ; platform_python_implementation == "CPython" and sys_platform == "win32" and python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
click-didyoumean==0.3.1 ; python_version >= "3.11" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
django-auditlog==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
django-celery-results==2.5.1 ; python_version >= "3.11" and python_version < "4.0"
django-cors-headers==4.4.0 ; python_version >= "3.11" and python_version < "4.0"
django==5.0.6 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework-simplejwt==5.3.1 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.15.2 ; python_version >= "3.11" and python_version < "4.0"
drf-spectacular==0.27.2 ; python_version >= "3.11" and python_version < "4.0"
drf-standardized-errors==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
gevent==24.2.1 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.0.3 ; platform_python_implementation == "CPython" and python_version >= "3.11" and python_version < "4.0"
gunicorn==22.0.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.7 ; python_version >= "3.11" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.11" and python_version < "4.0"
jsonschema==4.22.0 ; python_version >= "3.11" and python_version < "4.0"
kombu==5.3.7 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.11" and python_version < "4.0"
prompt-toolkit==3.0.47 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.22 ; platform_python_implementation == "CPython" and sys_platform == "win32" and python_version >= "3.11" and python_version < "4.0"
pyjwt==2.8.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
referencing==0.35.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rpds-py==0.18.1 ; python_version >= "3.11" and python_version < "4.0"
setuptools==70.1.1 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.5.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.11" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.11" and python_version < "4.0"
vine==5.1.0 ; python_version >= "3.11" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.11" and python_version < "4.0"
zope-event==5.0 ; python_version >= "3.11" and python_version < "4.0"
zope-interface==6.4.post2 ; python_version >= "3.11" and python_version < "4.0"
